+++ src/main/java/insane96mcp/iguanatweaksreborn/module/sleeprespawn/SleepingEffects.java
@@ -4,12 +4,10 @@ import insane96mcp.iguanatweaksreborn.InsaneSurvivalOverhaul;
 import insane96mcp.iguanatweaksreborn.data.ISOMobEffectInstance;
 import insane96mcp.iguanatweaksreborn.module.Modules;
 import insane96mcp.iguanatweaksreborn.module.sleeprespawn.tiredness.Tiredness;
-import insane96mcp.iguanatweaksreborn.module.sleeprespawn.tiredness.TirednessHandler;
 import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.base.*;
 import insane96mcp.insanelib.base.config.Config;
 import net.minecraft.network.chat.Component;
-import net.minecraft.world.effect.MobEffects;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.food.FoodData;
@@ -17,35 +15,36 @@ import net.minecraftforge.event.entity.player.PlayerSleepInBedEvent;
 import net.minecraftforge.event.level.SleepFinishedTimeEvent;
 import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
+import net.minecraftforge.fml.ModList;
 
 import java.util.ArrayList;
 import java.util.List;
 
 @Label(name = "Sleeping Effects", description = "Prevents the player from sleeping if has not enough Hunger and gives him effects on wake up. Effects on wake up are controlled via json in this feature's folder")
-@LoadFeature(module = Modules.Ids.SLEEP_RESPAWN, enabledByDefault = false)
+@LoadFeature(module = Modules.Ids.SLEEP_RESPAWN)
 public class SleepingEffects extends JsonFeature {
 
 	public static final List<ISOMobEffectInstance> EFFECTS_ON_WAKE_UP_DEFAULT = List.of(
-			new ISOMobEffectInstance.Builder(MobEffects.MOVEMENT_SLOWDOWN, 400).build(),
+			/*new ISOMobEffectInstance.Builder(MobEffects.MOVEMENT_SLOWDOWN, 400).build(),
 			new ISOMobEffectInstance.Builder(MobEffects.WEAKNESS, 300).setAmplifier(1).build(),
 			new ISOMobEffectInstance.Builder(MobEffects.DIG_SLOWDOWN, 300).setAmplifier(1).build(),
-			new ISOMobEffectInstance.Builder(MobEffects.REGENERATION, 900).build()
+			new ISOMobEffectInstance.Builder(MobEffects.REGENERATION, 600).build()*/
 	);
 	public static final ArrayList<ISOMobEffectInstance> effectsOnWakeUp = new ArrayList<>();
 	public static final String NO_FOOD_FOR_SLEEP = "iguanatweaksreborn.no_food_for_sleep";
 
 	@Config(min = 0, max = 20)
-	@Label(name = "Hunger Depleted on Wake Up", description = "How much the hunger bar is depleted when you wake up in the morning. Saturation is depleted before depleting hunger bar. Setting to 0 will disable this feature.")
+	@Label(name = "Hunger Depleted on Wake Up", description = "How much the hunger bar is depleted when you wake up in the morning. Setting to 0 will disable this feature.")
 	public static Integer hungerDepletedOnWakeUp = 15;
 	@Config
 	@Label(name = "No Sleep If Hungry", description = "If the player's hunger bar is below 'Hunger Depleted on Wake Up' he can't sleep.")
-	public static Boolean noSleepIfHungry = false;
+	public static Boolean noSleepIfHungry = true;
 	@Config
 	@Label(name = "No beneficial effect when hungry", description = "If the player has no hunger on wake up, beneficial effects are not applied.")
 	public static Boolean noBeneficialEffectWhenHungry = true;
-	@Config
-	@Label(name = "Dizzy when tired", description = "Apply the bad effects only when too tired")
-	public static Boolean dizzyWhenToTired = true;
+	@Config(min = -1, max = 255)
+	@Label(name = "Dizzy when tired", description = "Apply the bad effects only when the Tired effect is equal or above this amplifier. -1 to disable")
+	public static Integer dizzyWhenToTired = 3;
 
 	public SleepingEffects(Module module, boolean enabledByDefault, boolean canBeDisabled) {
 		super(module, enabledByDefault, canBeDisabled);
@@ -64,23 +63,28 @@ public class SleepingEffects extends JsonFeature {
 			return;
 
 		event.getLevel().players().stream().filter(LivingEntity::isSleeping).toList().forEach(player -> {
-			float tirednessOnWakeUp = TirednessHandler.getOnWakeUp(player);
-
-			FoodData foodData = player.getFoodData();
-			int hungerToDeplete = hungerDepletedOnWakeUp;
-			if (foodData.getSaturationLevel() > 0) {
-				float saturation = foodData.saturationLevel;
-				float saturationToDeplete = Math.min(hungerToDeplete, saturation);
-				foodData.setSaturation(saturation - saturationToDeplete);
-				hungerToDeplete -= saturationToDeplete;
+			if (!ModList.get().isLoaded("nohunger")) {
+				FoodData foodData = player.getFoodData();
+				int hungerToDeplete = hungerDepletedOnWakeUp;
+				/*if (foodData.getSaturationLevel() > 0) {
+					float saturation = foodData.saturationLevel;
+					int saturationToDeplete = (int) Math.min(hungerToDeplete, saturation);
+					foodData.setSaturation(saturation - saturationToDeplete);
+					hungerToDeplete -= saturationToDeplete;
+				}*/
+				if (hungerToDeplete > 0)
+					foodData.setFoodLevel(foodData.foodLevel - Math.min(hungerToDeplete, foodData.foodLevel));
 			}
-			if (hungerToDeplete > 0)
-				foodData.setFoodLevel(foodData.foodLevel - Math.min(hungerToDeplete, foodData.foodLevel));
 
 			for (ISOMobEffectInstance mobEffectInstance : effectsOnWakeUp) {
-				if (noBeneficialEffectWhenHungry && mobEffectInstance.effect.get().isBeneficial() && player.getFoodData().getFoodLevel() <= 0)
+				if (noBeneficialEffectWhenHungry
+						&& mobEffectInstance.effect.get().isBeneficial()
+						&& player.getFoodData().getFoodLevel() <= 0)
 					continue;
-				if (dizzyWhenToTired && Feature.isEnabled(Tiredness.class) && tirednessOnWakeUp == 0f && !mobEffectInstance.effect.get().isBeneficial())
+				if (dizzyWhenToTired > -1
+						&& !mobEffectInstance.effect.get().isBeneficial()
+						&& Feature.isEnabled(Tiredness.class)
+						&& (player.getEffect(Tiredness.TIRED.get()) == null || player.getEffect(Tiredness.TIRED.get()).getAmplifier() < dizzyWhenToTired))
 					continue;
 				player.addEffect(mobEffectInstance.getMobEffectInstance());
 			}